Made some changes to the set fitting routines.
I found a bug in the treatment of the error values in the creation of a new set by fitting existing data.  This led to issues with saving plots in that case, as the error arrays were smaller than the data arrays.  A minor change was made to the range of the fitted function.