Changes to closing windows.
The code has internal variables for many windows that need to be set to None upon exit so that the window will be recreated the next time, and to prevent having duplicate windows at one time.  There was an issue with closing the windows in that the internal variable was not set correctly back to None.  The code here should fix this issue.  Some duplicate code was found in the process, and removed.