Codes run through Spyder.
All of my codes (therefore not mpfit.py and mpfitexpr.py) were run through the Spyder code and various changes were made in response to the code analysis.  A few bugs we found and fixed in the process, but most of the changes were "cosmetic".  These need testing to make sure that the changes have not caused indirect problems.