Changed the mouse-click zoom function.
There was an issue with the mouse-click zoom function, it was reacting to clicks very close together (which were not intended for zooming) and there was an issue with the storage of the mouse click positions for the purpose of zooming.  These have been fixed.  One cannot zoom to less than 1% of the current x or y axis range in one go.