Fix point matching to scale by axis range
The distance calculation can have problems in the case where 1 axis has a very different magnitude than the other, because in that case the distances along the axis with the much larger range dominates the distance calculation.  A fix has been put in to normalize the distances in terms of the x and y range so the two axes are on equal footing no matter how different the magnitudes of the values may be.